refactor: 입력 예외 처리 클래스, 패키지명 수정
